requirements.txt
Flask>=3.1.0
Pillow>=11.0.0
piexif>=1.1.3
